+++ CHANGELOG.md
@@ -46,7 +46,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Collapsible side bar.
 - [frontend] - Add confirmation box on thread deletion.
 - Feedback for HIL tool call refusal.
+- [frontend] - Per message tool toggle.
 - [frontend] - Possibility to dynamically chose available tools.
+- Displaying tool metadata
 
 ### Fixed
 - [frontend] - Parallel execution of tools
@@ -54,7 +56,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Query size limiter.
 - Fix /threads/messages endpoint.
 - [frontend] - remove first message delay.
+- [frontend] - improve loading UI on first message.
+- [frontend] - MarkDown rendering.
 - [frontend] - dark mode flickering.
+- [frontend] - Concatenate base url and endpoint correctly.
 
 ## [0.3.3] - 30.10.2024